creates OnBoardedApp
