+++ apps/gp2-frontend/src/AuthenticatedApp.tsx
@@ -1,38 +1,19 @@
 import { Auth0, gp2 as gp2Auth } from '@asap-hub/auth';
 import { Frame } from '@asap-hub/frontend-utils';
-import { BasicLayout, Layout } from '@asap-hub/gp2-components';
+import { BasicLayout } from '@asap-hub/gp2-components';
 import { Loading, NotFoundPage } from '@asap-hub/react-components';
 import { useAuth0GP2, useCurrentUserGP2 } from '@asap-hub/react-context';
-import { gp2 as gp2Route } from '@asap-hub/routing';
 import { FC, lazy, useEffect } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { RecoilRoot, useRecoilState, useResetRecoilState } from 'recoil';
 import { auth0State } from './auth/state';
-import { useUserById } from './users/state';
 
-const {
-  workingGroups: workingGroupsRoute,
-  users: usersRoute,
-  projects: projectsRoute,
-} = gp2Route;
-const loadDashboard = () =>
-  import(/* webpackChunkName: "dashboard" */ './dashboard/Dashboard');
-
-const loadWorkingGroups = () =>
-  import(/* webpackChunkName: "working-groups" */ './working-groups/Routes');
-
-const loadProjects = () =>
-  import(/* webpackChunkName: "projects" */ './projects/Routes');
-
-const loadUsers = () =>
-  import(/* webpackChunkName: "users" */ './users/Routes');
+const loadOnboardedApp = () =>
+  import(/* webpackChunkName: "onboarded-app" */ './OnboardedApp');
 const loadOnboarding = () =>
   import(/* webpackChunkName: "onboarding" */ './onboarding/Routes');
 
-const Dashboard = lazy(loadDashboard);
-const WorkingGroups = lazy(loadWorkingGroups);
-const Projects = lazy(loadProjects);
-const Users = lazy(loadUsers);
+const OnboardedApp = lazy(loadOnboardedApp);
 const Onboarding = lazy(loadOnboarding);
 
 const AuthenticatedApp: FC<Record<string, never>> = () => {
@@ -51,60 +32,20 @@ const AuthenticatedApp: FC<Record<string, never>> = () => {
 
   useEffect(() => {
     // order by the likelyhood of user navigating there
-    user?.onboarded
-      ? loadDashboard()
-          .then(loadUsers)
-          .then(loadWorkingGroups)
-          .then(loadProjects)
-      : loadOnboarding();
+    user?.onboarded ? loadOnboardedApp() : loadOnboarding();
   }, [user?.onboarded]);
-
-  const { projects = [], workingGroups = [] } =
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    useUserById(user!.id) || {};
 
   if (!user || !recoilAuth0) {
     return <Loading />;
   }
 
-  if (!user.onboarded) {
-    return (
-      <BasicLayout>
-        <Switch>
-          <Route path={path}>
-            <Onboarding />
-          </Route>
-          <Route>
-            <Frame title="Not Found">
-              <NotFoundPage />
-            </Frame>
-          </Route>
-        </Switch>
-      </BasicLayout>
-    );
-  }
-  return (
-    <Layout projects={projects} workingGroups={workingGroups}>
+  return user.onboarded ? (
+    <OnboardedApp />
+  ) : (
+    <BasicLayout>
       <Switch>
-        <Route exact path={path}>
-          <Frame title="Dashboard">
-            <Dashboard />
-          </Frame>
-        </Route>
-        <Route path={usersRoute.template}>
-          <Frame title="Users">
-            <Users />
-          </Frame>
-        </Route>
-        <Route path={workingGroupsRoute.template}>
-          <Frame title="Working Groups">
-            <WorkingGroups />
-          </Frame>
-        </Route>
-        <Route path={projectsRoute.template}>
-          <Frame title="Projects">
-            <Projects />
-          </Frame>
+        <Route path={path}>
+          <Onboarding />
         </Route>
         <Route>
           <Frame title="Not Found">
@@ -112,7 +53,7 @@ const AuthenticatedApp: FC<Record<string, never>> = () => {
           </Frame>
         </Route>
       </Switch>
-    </Layout>
+    </BasicLayout>
   );
 };
